+++ src/pages/write.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 import PostEditor from "src/components/write/PostEditor";
-import dynamic from "next/dynamic";
